Allow overflowing restaurant names to wrap
Previously, long restaurant names would be cut off with an ellipsis.
Now, they wrap to the next line (and have a smaller font-size to avoid
too much wrapping). This means that sometimes, restaurant names will
wrap in an ugly way, but overall it improves the readability of the
restaurant names.